bug fixes for xs:decimal, xs:integer and corresponding stuff
darcs-hash:20060910101725-10a17-97fde823a4923dfdc9408ea01a2603d2c1e10bf0.gz

+++ kernel/tr/executor/xqops/PPReturn.cpp
@@ -180,7 +180,7 @@ void PPReturn::next(tuple &t, var_dsc dsc, var_c_id id)
     if (p.svc->at(id))
     {
         p.svc->at(id) = false;
-        t.copy(dsc == pos_dsc ? tuple_cell::atomic(pos)
+        t.copy(dsc == pos_dsc ? tuple_cell::atomic((__int64)pos)
                               : source.cells[p.tuple_pos]);
     }
 	else
